Fix reviews embed

+++ src/review/views.py
@@ -7,6 +7,7 @@ from rest_framework.pagination import PageNumberPagination
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from user.auth_methods.auth0_authentication import Auth0JWTAuthentication
+from user.models import CustomUser as User
 
 from .controllers import (
     create_review,
@@ -15,7 +16,11 @@ from .controllers import (
     get_reviews_for_entity,
     update_review,
 )
-from .serializers import PaginatedReviewResponseSerializer, ReviewSerializer
+from .serializers import (
+    EntityReviewSerializer,
+    PaginatedEntityReviewResponseSerializer,
+    ReviewSerializer,
+)
 
 
 class ReviewViewSet(viewsets.ViewSet):
@@ -41,22 +46,24 @@ class ReviewViewSet(viewsets.ViewSet):
         responses={
             200: openapi.Response(
                 description="Paginated list of reviews",
-                schema=PaginatedReviewResponseSerializer,
+                schema=PaginatedEntityReviewResponseSerializer,
             )
         },
         security=[{"Auth0JWT": []}],
     )
     def list(self, request, entity_pk=None):
-        entity = get_entity_by_id(entity_pk)
-        reviews = get_reviews_for_entity(entity)
+        entity = get_entity_by_id(request.user, entity_pk, is_service_id=False)
+        reviews = get_reviews_for_entity(entity).order_by("-created_at")
 
         if request.query_params.get("all") == "true":
-            serializer = ReviewSerializer(reviews, many=True)
+            serializer = EntityReviewSerializer(reviews, many=True)
             return Response(serializer.data)
 
         paginator = PageNumberPagination()
         paginated_reviews = paginator.paginate_queryset(reviews, request)
-        serializer = ReviewSerializer(paginated_reviews, many=True)
+        serializer = EntityReviewSerializer(
+            paginated_reviews, many=True, context={"request": request}
+        )
         return paginator.get_paginated_response(serializer.data)
 
     @swagger_auto_schema(
@@ -65,6 +72,7 @@ class ReviewViewSet(viewsets.ViewSet):
     )
     def retrieve(self, _, pk=None, entity_pk=None):
         review = get_review_by_id(pk)
+
         if str(review.entity.id) != str(entity_pk):
             raise NotFound("Review not found for the specified entity")
         serializer = ReviewSerializer(review)
@@ -76,10 +84,11 @@ class ReviewViewSet(viewsets.ViewSet):
         security=[{"Auth0JWT": []}],
     )
     def create(self, request, entity_pk=None):
-        entity = get_entity_by_id(entity_pk)
+        entity = get_entity_by_id(request.user, entity_pk, is_service_id=False)
+
         serializer = ReviewSerializer(data=request.data)
         if serializer.is_valid():
-            review = create_review(entity, serializer.validated_data)
+            review = create_review(entity, request.user, serializer.validated_data)
             return Response(
                 ReviewSerializer(review).data, status=status.HTTP_201_CREATED
             )
@@ -94,6 +103,9 @@ class ReviewViewSet(viewsets.ViewSet):
         review = get_review_by_id(pk)
         if str(review.entity.id) != str(entity_pk):
             raise NotFound("Review not found for the specified entity")
+        if User.objects.get(id=request.user.id) != review.user:
+            raise NotFound("You can only update your own reviews")
+
         serializer = ReviewSerializer(review, data=request.data)
         if serializer.is_valid():
             review = update_review(review, serializer.validated_data)
@@ -104,9 +116,12 @@ class ReviewViewSet(viewsets.ViewSet):
         responses={204: "No Content"},
         security=[{"Auth0JWT": []}],
     )
-    def destroy(self, _, pk=None, entity_pk=None):
+    def destroy(self, request, pk=None, entity_pk=None):
         review = get_review_by_id(pk)
         if str(review.entity.id) != str(entity_pk):
             raise NotFound("Review not found for the specified entity")
+        if User.objects.get(id=request.user.id) != review.user:
+            raise NotFound("You can only update your own reviews")
+
         delete_review(review)
         return Response(status=status.HTTP_204_NO_CONTENT)
